requirements.txt
# Core dependencies
mcp>=0.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.1
pytest-mock>=3.11.1
coverage>=7.3.0
hypothesis>=6.82.6
click>=8.1.0
requests>=2.31.0
rich>=13.7.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-prometheus>=0.52b1
opentelemetry-exporter-otlp-proto-http>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
prometheus-client>=0.19.0
psutil>=5.9.0
bandit>=1.7.8
ruff>=0.3.0
PyYAML>=6.0.1
tree-sitter>=0.20.1

# LLM dependencies (for llm_server.py)
torch>=2.2.0
transformers>=4.38.0
anthropic>=0.18.0
openai>=1.12.0

# Added from the code block
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic>=1.8.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
requests>=2.26.0
aiohttp>=3.8.0
opentelemetry-api>=1.0.0
opentelemetry-sdk>=1.0.0
opentelemetry-instrumentation-fastapi>=0.0.0
opentelemetry-exporter-otlp>=1.0.0
prometheus-client>=0.11.0
python-json-logger>=2.0.0
pyyaml>=5.4.1
anthropic>=0.3.0
openai>=0.27.0
torch>=1.9.0
transformers>=4.11.0
digitalocean>=1.17.0

# Monitoring and load testing
prometheus-client>=0.21.1
locust>=2.33.2

# Optional dependencies
opentelemetry-api>=1.0.0
opentelemetry-sdk>=1.0.0
opentelemetry-instrumentation-fastapi>=0.0.0
opentelemetry-exporter-otlp>=1.0.0
opentelemetry-exporter-jaeger>=1.0.0
opentelemetry-exporter-zipkin>=1.0.0
opentelemetry-exporter-prometheus>=1.0.0
opentelemetry-instrumentation-requests>=0.0.0
opentelemetry-instrumentation-sqlalchemy>=0.0.0
opentelemetry-instrumentation-psycopg2>=0.0.0
opentelemetry-instrumentation-pymongo>=0.0.0
opentelemetry-instrumentation-redis>=0.0.0
opentelemetry-instrumentation-grpc>=0.0.0
opentelemetry-instrumentation-httpx>=0.0.0
opentelemetry-instrumentation-aiohttp-client>=0.0.0
opentelemetry-instrumentation-asgi>=0.0.0
opentelemetry-instrumentation-wsgi>=0.0.0
opentelemetry-instrumentation-django>=0.0.0
opentelemetry-instrumentation-flask>=0.0.0
opentelemetry-instrumentation-tornado>=0.0.0
opentelemetry-instrumentation-sqlite3>=0.0.0
opentelemetry-instrumentation-mysql>=0.0.0
opentelemetry-instrumentation-postgresql>=0.0.0
opentelemetry-instrumentation-sqlite3>=0.0.0
opentelemetry-instrumentation-elasticsearch>=0.0.0
opentelemetry-instrumentation-kafka>=0.0.0
opentelemetry-instrumentation-rabbitmq>=0.0.0
opentelemetry-instrumentation-celery>=0.0.0
opentelemetry-instrumentation-django-celery>=0.0.0
opentelemetry-instrumentation-flask-celery>=0.0.0
opentelemetry-instrumentation-fastapi-celery>=0.0.0
opentelemetry-instrumentation-asyncio>=0.0.0
opentelemetry-instrumentation-aio-pika>=0.0.0
opentelemetry-instrumentation-aio-redis>=0.0.0
opentelemetry-instrumentation-aio-zipkin>=0.0.0
opentelemetry-instrumentation-aio-jaeger>=0.0.0
opentelemetry-instrumentation-aio-otlp>=0.0.0
opentelemetry-instrumentation-aio-prometheus>=0.0.0
opentelemetry-instrumentation-aio-grpc>=0.0.0
opentelemetry-instrumentation-aio-http>=0.0.0
opentelemetry-instrumentation-aio-sqlalchemy>=0.0.0
opentelemetry-instrumentation-aio-pymongo>=0.0.0
opentelemetry-instrumentation-aio-elasticsearch>=0.0.0
opentelemetry-instrumentation-aio-kafka>=0.0.0
opentelemetry-instrumentation-aio-rabbitmq>=0.0.0
opentelemetry-instrumentation-aio-celery>=0.0.0
opentelemetry-instrumentation-aio-django-celery>=0.0.0
opentelemetry-instrumentation-aio-flask-celery>=0.0.0
opentelemetry-instrumentation-aio-fastapi-celery>=0.0.0
opentelemetry-instrumentation-aio-asyncio>=0.0.0 